internal change to data system
embedded it in main react tree; hoping to modularize data system. in
particular, want to be able to store data lower in tree,
e.g., zoom and position data in SvgScene, that can be added to
atomStoreAtom, which will trigger rerender of MainDataComp, and so
make saving and restoring for data that is kept to lower part of tree.